# routes/movies.js
import express from 'express';

import MoviesService from '../services/movies.js';
import validationHandler from '../utils/middlewares/validationHandler.js';
import {
  movieIdSchema,
  createMovieSchema,
  updateMovieSchema,
} from '../utils/schemas/movies.js';

const moviesApi = (app) => {
  const router = express.Router();
  app.use('/movies', router);

  const moviesService = new MoviesService();

  router.get('/', async (req, res, next) => {
    const { tags } = req.query;

    try {
      const movies = await moviesService.getMovies(tags);

      res.status(200).json({
        data: movies,
        message: 'movies listed',
      });
    } catch (error) {
      next(error);
    }
  });

  router.get(
    '/:movieId',
    validationHandler({ movieId: movieIdSchema }, 'params'),
    async (req, res, next) => {
      const { movieId } = req.params;

      try {
        const movie = await moviesService.getMovie(movieId);

        if (movie) {
          res.status(200).json({
            data: movie,
            message: 'movie retrieved',
          });
        } else {
          res.status(404).json({
            message: 'movie not found',
          });
        }
      } catch (error) {
        next(error);
      }
    }
  );

  router.post(
    '/',
    validationHandler(createMovieSchema),
    async (req, res, next) => {
      const { body: movie } = req;

      try {
        const createdMovieId = await moviesService.createMovie(movie);

        res.status(201).json({
          data: createdMovieId,
          message: 'movie created',
        });
      } catch (error) {
        next(error);
      }
    }
  );

  router.put(
    '/:movieId',
    validationHandler({ movieId: movieIdSchema }, 'params'),
    validationHandler(updateMovieSchema),
    async (req, res, next) => {
      const { movieId } = req.params;
      const { body: movie } = req;

      try {
        const updatedMovieId = await moviesService.updateMovie(movieId, movie);

        res.status(200).json({
          data: updatedMovieId,
          message: 'movie updated',
        });
      } catch (error) {
        next(error);
      }
    }
  );

  router.delete(
    '/:movieId',
    validationHandler({ movieId: movieIdSchema }, 'params'),
    async (req, res, next) => {
      const { movieId } = req.params;

      try {
        const deletedMovieId = await moviesService.deleteMovie(movieId);

        res.status(200).json({
          data: deletedMovieId,
          message: 'movie deleted',
        });
      } catch (error) {
        next(error);
      }
    }
  );
};

export default moviesApi;